implementation de da

+++ include/arch/address.h
@@ -8,5 +8,5 @@ int display_one_addr(ARCH, char*);
 int display_addr_to_addr(ARCH, char*);
 int display_bytes_from_addr(ARCH, char*);
 int display_range_addr(ARCH, uint, int);
-int get_addr(ARCH, uint, uint*);
+int get_addr(ARCH, uint, uchar*);
 #endif
